fix(Series): fix null in series

+++ frappe/database/mariadb/framework_mariadb.sql
@@ -239,7 +239,7 @@ CREATE TABLE `tabDocType` (
 DROP TABLE IF EXISTS `tabSeries`;
 CREATE TABLE `tabSeries` (
   `name` varchar(100),
-  `current` int(10) NOT NULL DEFAULT 0,
+  `current` bigint(20) NOT NULL DEFAULT 0,
   PRIMARY KEY(`name`)
 ) ENGINE=InnoDB ROW_FORMAT=DYNAMIC CHARACTER SET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
